refactor: log setting 페이지 구조 변경
